+++ bee-identity/src/keypair.rs
@@ -1,7 +1,0 @@
-// Copyright 2022 IOTA Stiftung
-// SPDX-License-Identifier: Apache-2.0
-
-//! TODO
-
-#[doc(inline)]
-pub use libp2p_core::identity::ed25519::{Keypair, PublicKey, SecretKey};