+++ bee-identity/CHANGELOG.md
@@ -31,4 +31,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-First release of the bee-identity.
+- `Identity` type;
+- `pem_file` module;
+